Struct streamer::platform::ProfileBuilder

pub struct ProfileBuilder { /* fields omitted */ }

A builder of platform profiles.

Methods

impl ProfileBuilder

fn new(element_count: usize, time_step: f64, fill: Vec<f64>) -> ProfileBuilder

Create a builder.

fn push(&mut self, element_id: usize, time: f64, time_step: f64, data: &[f64])

Add data to a particular processing element starting from a particular time moment.

fn pull(&mut self, time: f64) -> Profile

Advance time and return the data accumulated since the previous call.