Trait streamer::platform::Platform

pub trait Platform {
    type Data;
    fn elements(&self) -> &[Element];

    fn next(&mut self, _: f64) -> Result<Self::Data>;

    fn push(&mut self, _: &Job, start: f64, mapping: &Mapping) -> Result<()>;
}

A multiprocessor platform.

Associated Types

type Data

The data produced by the platform.

Required Methods

fn elements(&self) -> &[Element]

Return the processing elements.

fn next(&mut self, _: f64) -> Result<Self::Data>

Advance time and return the data accumulated since the previous call.

fn push(&mut self, _: &Job, start: f64, mapping: &Mapping) -> Result<()>

Account for a scheduling decision taken with respect to a job.